backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.0.1
django-filter==22.1
django-recurrence==1.11.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
djangorestframework-simplejwt==5.2.2
django-phone-field==1.8.1